% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example graph from Wikipedia}
\format{
A data.frame with 18 rows and 3 variables:
\describe{
\item{v1}{Start node of the edge.}
\item{v2}{End node of the edge.}
\item{w}{Weight of the edge.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
The example graph used on the Wikipedia page for Dijkstra's algorithm.
}
\keyword{datasets}
