% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric scalar representing the first integer.}

\item{b}{A numeric scalar representing the second integer.}
}
\value{
A numeric scalar representing the greatest common divisor (GCD) of a and b.
}
\description{
This function computes the greatest common divisor (GCD) of two integers using the Euclidean algorithm.
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
